+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "bluerain-plugin-material-ui",
+  "name": "@blueeast/bluerain-plugin-material-ui",
   "version": "0.5.0",
   "description": "A BlueRain Plugin that integrates material-ui",
   "keywords": [],
