+++ test/bluerain-plugin-material-ui.test.ts
@@ -1,10 +1,33 @@
-import DummyClass from "../src/bluerain-plugin-material-ui"
+import BR, { Plugin } from '@blueeast/bluerain-os';
+// tslint:disable-next-line:no-implicit-dependencies
+import React from 'react';
+import MaterialUiPlugin from '../src/index';
 
-/**
- * Dummy test
- */
-describe("Dummy test", () => {
-  it("works if true is truthy", () => {
-    expect(true).toBeTruthy()
-  })
+beforeEach(() => {
+	document.body.innerHTML = '<div class="app-container">' + '</div>';
+});
 
+describe('material-ui Plugin Test', () => {
+	it('should register plugin successfully', () => {
+		BR.Plugins.register(MaterialUiPlugin);
+		const plugin = BR.Plugins.get(MaterialUiPlugin.slug);
+		expect(plugin).toBeTruthy();
+	});
+
+	it('should   return  undefined theme ', () => {
+		BR.boot();
+		expect(BR.Configs.get('theme')).toBeUndefined();
+	});
+
+	it('should   return defined theme', () => {
+		const theme = {
+			color: 'black',
+			isRtl: ''
+		};
+
+		BR.Configs.set('theme', theme);
+		BR.boot();
+
+		expect(BR.Configs.get('theme')).toEqual(theme);
+	});
+});
